+++ src/messages/response.py
@@ -11,3 +11,4 @@ class ReservationResponse(BaseMessage):
 
     success: bool
     user_id: str
+    reservation_id: str