+++ CHANGELOG.md
@@ -10,6 +10,14 @@ Format according to the [keepachangelog](https://keepachangelog.com/en/1.1.0/).
 - `ParkingAgent._available_parking_spots` is a list now - it represents next 24 hours of parking spots availability.
 
 
+## [0.1.4] - 2024-12-04
+### Added
+- `User` agent with main methods for creating and modifying a reservation
+
+### Changed
+- `reservation_id` attribute to `ReservationResponse` message
+
+
 ## [0.1.3] - 2024-12-01
 ### Changed
 - We have to make assumption that every location is covered with RegionalCoordinator.
@@ -18,6 +26,7 @@ Format according to the [keepachangelog](https://keepachangelog.com/en/1.1.0/).
 ## [0.1.2] - 2024-11-30
 ### Removed
 - `hello` agent development directory.
+
 
 ## [0.1.1] - 2024-11-29
 ### Changed
@@ -31,4 +40,4 @@ Format according to the [keepachangelog](https://keepachangelog.com/en/1.1.0/).
 - Proposed using `dataclass` for message bodies in `src/messages`, loaded with:
   ```python
   message_body=NazwaKlasy(**json.loads(message.body))
-
+  ```